Centralised data IO fixtures to conftest.py.
